firecitadel/requirements.txt
openai
tweepy
langchain
python-dotenv
